requirements.txt
# =========================================================================================
#  Copyright 2016 Community Information Online Consortium (CIOC) and KCL Software Solutions
#
#  Licensed under the Apache License, Version 2.0 (the "License");
#  you may not use this file except in compliance with the License.
#  You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
#  Unless required by applicable law or agreed to in writing, software
#  distributed under the License is distributed on an "AS IS" BASIS,
#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#  See the License for the specific language governing permissions and
#  limitations under the License.
# =========================================================================================

pyramid==1.4.9
SQLAlchemy==0.7.10
transaction==1.4.4
requests==2.9.1
pyramid_tm==0.12.1
zope.sqlalchemy
https://github.com/formencode/formencode/archive/master.zip#egg=formencode
paste
pyramid_beaker==0.8
backports.ssl_match_hostname
PyCrypto==2.6.1
webhelpers2
https://github.com/Pylons/pyramid_simpleform/archive/master.zip#egg=pyramid_simpleform
pyramid_exclog
Babel
wincertstore
apscheduler==2.0.2
